Refactor: Move steady-state code to forward/backward problem
Move steady-state code related to the forward problem to forward_problem.{h,cpp}
and steady-state code related to the backward problem to backward_problem.{h,cpp}
to avoid approaching circular include issues.

No further changes.
