+++ include/amici/backwardproblem.h
@@ -136,6 +136,134 @@ class EventHandlingBwdSimulator {
     realtype t_{0};
 };
 
+/**
+ * @brief The SteadyStateBackwardProblem class computes the adjoint state and
+ * quadratures for pre- or post-equilibration.
+ */
+class SteadyStateBackwardProblem {
+  public:
+    /**
+     * @brief SteadyStateBackwardProblem ctor
+     * @param solver
+     * @param model
+     * @param final_state Final state from pre/post-equilibration forward
+     * problem
+     * @param ws Workspace for backward simulation
+     */
+    SteadyStateBackwardProblem(
+        Solver const& solver, Model& model, SimulationState& final_state,
+        gsl::not_null<BwdSimWorkspace*> ws
+    );
+
+    /**
+     * @brief Compute the gradient via adjoint steady-state sensitivities.
+     *
+     * Integrates over the adjoint state backward in time by solving a linear
+     * system of equations, which gives the analytical solution.
+     *
+     * Expects the workspace `ws_` to be initialized.
+     *
+     * The results will be written to the workspace `ws_`.
+     *
+     * @param t0 Initial time for the steady state simulation.
+     */
+    void run(realtype t0);
+
+    /**
+     * @brief Get the CPU time taken to solve the backward problem.
+     * @return The CPU time in milliseconds.
+     */
+    [[nodiscard]] double getCPUTimeB() const { return cpu_timeB_; }
+
+    /**
+     * @brief Get the number of steps taken to find the steady state in the
+     * adjoint case.
+     * @return Number of steps.
+     */
+    [[nodiscard]] int getNumStepsB() const { return numstepsB_; }
+
+    /**
+     * @brief Return the adjoint state
+     *
+     * Accessible only after run() has been called and before ws_ is used
+     * elsewhere.
+     *
+     * @return xB adjoint state
+     */
+    [[nodiscard]] AmiVector const& getAdjointState() const;
+
+    /**
+     * @brief Get the adjoint quadratures (xQB).
+     *
+     * Accessible only after run() has been called and before ws_ is used
+     * elsewhere.
+
+     * @return xQB
+     */
+    [[nodiscard]] AmiVector const& getAdjointQuadrature() const;
+
+    /**
+     * @brief Accessor for has_quadrature_
+     * @return has_quadrature_
+     */
+    [[nodiscard]] bool hasQuadrature() const { return has_quadrature_; }
+
+  private:
+    /**
+     * @brief Launch backward simulation if Newton solver or linear system solve
+     * fail or are disabled.
+     * @param solver Solver instance.
+     */
+    void run_simulation(Solver const& solver);
+
+    /**
+     * @brief Compute quadratures in adjoint mode
+     * @param t0 Initial time for the steady state simulation.
+     */
+    void compute_steady_state_quadrature(realtype t0);
+
+    /**
+     * @brief Compute the quadrature in steady state backward mode by
+     * solving the linear system defined by the backward Jacobian.
+     */
+    void compute_quadrature_by_lin_solve();
+
+    /**
+     * @brief Computes the quadrature in steady state backward mode by
+     * numerical integration of xB forward in time.
+     * @param t0 Initial time for the steady state simulation.
+     */
+    void compute_quadrature_by_simulation(realtype t0);
+
+    /** CPU time for solving the backward problem (milliseconds) */
+    double cpu_timeB_{0.0};
+
+    /** flag indicating whether backward mode was run */
+    bool has_quadrature_{false};
+
+    /** The employed number of backward steps */
+    int numstepsB_{0};
+
+    /** integral over adjoint state vector */
+    AmiVector xQ_;
+
+    /** Final state from pre/post-equilibration forward problem */
+    SimulationState& final_state_;
+
+    /** Newton solver */
+    NewtonSolver newton_solver_;
+
+    /**
+     * Whether the Newton step should be used instead of xdot for convergence
+     * checks during simulation.
+     */
+    bool newton_step_conv_{false};
+
+    Model* model_{nullptr};
+    Solver const* solver_{nullptr};
+    BwdSimWorkspace* ws_{nullptr};
+};
+
 //!  class to solve backwards problems.
 /*!
   solves the backwards problem for adjoint sensitivity analysis and handles
