+++ src/forwardproblem.cpp
@@ -13,6 +13,8 @@
 
 namespace amici {
 
+constexpr realtype conv_thresh = 1.0;
+
 /**
  * @brief Check if the next timepoint is too close to the current timepoint.
  *
@@ -701,6 +703,472 @@ std::vector<int> compute_nroots(
         }
     }
     return nroots;
+}
+
+/**
+ * @brief Assemble the error message to be thrown according to steady state
+ * computation status.
+ * @param errorString The error string to append to.
+ * @param status Status of the steady state computation.
+ */
+static void
+writeErrorString(std::string& errorString, SteadyStateStatus status) {
+    switch (status) {
+    case SteadyStateStatus::failed_too_long_simulation:
+        errorString.append(": System could not be equilibrated.");
+        break;
+    case SteadyStateStatus::failed_damping:
+        errorString.append(": Damping factor reached lower bound.");
+        break;
+    case SteadyStateStatus::failed_factorization:
+        errorString.append(": RHS could not be factorized.");
+        break;
+    case SteadyStateStatus::failed_convergence:
+        errorString.append(": No convergence was achieved.");
+        break;
+    default:
+        errorString.append(".");
+        break;
+    }
+}
+
+SteadystateProblem::SteadystateProblem(
+    FwdSimWorkspace* ws, Solver const& solver, Model& model
+)
+    : ws_(ws)
+    , wrms_computer_x_(
+          model.nx_solver, solver.getSunContext(),
+          solver.getAbsoluteToleranceSteadyState(),
+          solver.getRelativeToleranceSteadyState(),
+          AmiVector(model.get_steadystate_mask(), solver.getSunContext())
+      )
+    , final_state_(
+          {.t = INFINITY,
+           .x = AmiVector(model.nx_solver, solver.getSunContext()),
+           .dx = AmiVector(model.nx_solver, solver.getSunContext()),
+           .sx = AmiVectorArray(
+               model.nx_solver, model.nplist(), solver.getSunContext()
+           ),
+           .state = model.getModelState()}
+      )
+    , newton_solver_(
+          NewtonSolver(model, solver.getLinearSolver(), solver.getSunContext())
+      )
+    , newtons_method_(
+          &model, solver.getSunContext(), &newton_solver_,
+          solver.getNewtonDampingFactorMode(),
+          solver.getNewtonDampingFactorLowerBound(), solver.getNewtonMaxSteps(),
+          solver.getNewtonStepSteadyStateCheck()
+      )
+    , newton_step_conv_(solver.getNewtonStepSteadyStateCheck())
+    , solver_(&solver) {
+    // Check for compatibility of options
+    if (solver.getSensitivityMethod() == SensitivityMethod::forward
+        && solver.getSensitivityMethodPreequilibration()
+               == SensitivityMethod::adjoint
+        && solver.getSensitivityOrder() > SensitivityOrder::none)
+        throw AmiException(
+            "Preequilibration using adjoint sensitivities "
+            "is not compatible with using forward "
+            "sensitivities during simulation"
+        );
+    if (solver.getSensitivityMethod() == SensitivityMethod::forward
+        && model.getSteadyStateComputationMode()
+               == SteadyStateComputationMode::newtonOnly
+        && model.getSteadyStateSensitivityMode()
+               == SteadyStateSensitivityMode::integrationOnly)
+        throw AmiException(
+            "For forward sensitivity analysis steady-state "
+            "computation mode 'newtonOnly' and steady-state "
+            "sensitivity mode 'integrationOnly' are not "
+            "compatible as numerical integration of the model "
+            "ODEs and corresponding forward sensitivities ODEs "
+            "is coupled"
+        );
+}
+
+void SteadystateProblem::workSteadyStateProblem(
+    Solver& solver, Model& model, int it, realtype t0
+) {
+    // Compute steady state, track computation time
+    CpuTimer cpu_timer;
+    ws_->t = t0;
+
+    // handle initial events
+    EventHandlingSimulator sim(&model, &solver, ws_, nullptr);
+    sim.handle_event(true, nullptr);
+
+    flagUpdatedState();
+    newton_solver_.reinitialize();
+    findSteadyState(model, it, t0);
+
+    // Check whether state sensitivities still need to be computed.
+    // Did we already compute forward sensitivities?
+    bool const forwardSensisAlreadyComputed
+        = solver.getSensitivityOrder() >= SensitivityOrder::first
+          && steady_state_status_[1] == SteadyStateStatus::success
+          && (model.getSteadyStateSensitivityMode()
+                  == SteadyStateSensitivityMode::integrationOnly
+              || model.getSteadyStateSensitivityMode()
+                     == SteadyStateSensitivityMode::integrateIfNewtonFails);
+    bool const simulationStartedInSteadystate
+        = steady_state_status_[0] == SteadyStateStatus::success
+          && numsteps_[0] == 0;
+    // Do we need forward sensis for postequilibration?
+    // Are we running in preequilibration (and hence create)?
+    bool const is_preequilibration = (it == -1);
+    // Do we need forward sensitivities for pre- or post-equilibration?
+    bool const needForwardSensisPosteq
+        = !is_preequilibration && !forwardSensisAlreadyComputed
+          && solver.getSensitivityOrder() >= SensitivityOrder::first
+          && solver.getSensitivityMethod() == SensitivityMethod::forward;
+    bool const needForwardSensisPreeq
+        = is_preequilibration && !forwardSensisAlreadyComputed
+          && solver.getSensitivityMethodPreequilibration()
+                 == SensitivityMethod::forward
+          && solver.getSensitivityOrder() >= SensitivityOrder::first;
+
+    // Do we need to do the linear system solve to get forward sensitivities?
+    if ((needForwardSensisPreeq || needForwardSensisPosteq)
+        && !simulationStartedInSteadystate) {
+        try {
+            // This might still fail, if the Jacobian is singular and
+            // simulation did not find a steady state.
+            newton_solver_.computeNewtonSensis(
+                ws_->sx, model, {ws_->t, ws_->x, ws_->dx}
+            );
+        } catch (NewtonFailure const&) {
+            throw AmiException(
+                "Steady state sensitivity computation failed due "
+                "to unsuccessful factorization of RHS Jacobian"
+            );
+        }
+    }
+    cpu_time_ = cpu_timer.elapsed_milliseconds();
+    final_state_.state = model.getModelState();
+    final_state_.t = ws_->t;
+    final_state_.x = ws_->x;
+    final_state_.dx = ws_->dx;
+    final_state_.sx = ws_->sx;
+}
+
+SteadyStateStatus SteadystateProblem::findSteadyStateBySimulation(
+    Model& model, int const it, realtype const t0
+) {
+    try {
+        // Preequilibration -> Create a new solver instance for simulation
+        // Postequilibration -> Solver was already created, use that one
+        if (it < 0) {
+            // TODO(performance): We should be able to avoid this clone
+            //   if we aren't using ASA in combination with events.
+            auto main_solver = solver_;
+            auto new_solver = std::unique_ptr<Solver>(main_solver->clone());
+            new_solver->logger = main_solver->logger;
+
+            // do we need sensitivities?
+            if (new_solver->getSensitivityMethodPreequilibration()
+                    == SensitivityMethod::forward
+                && new_solver->getSensitivityOrder() >= SensitivityOrder::first
+                && (model.getSteadyStateSensitivityMode()
+                        == SteadyStateSensitivityMode::integrationOnly
+                    || model.getSteadyStateSensitivityMode()
+                           == SteadyStateSensitivityMode::integrateIfNewtonFails
+                )) {
+                // need forward to compute sx0 for the pre/main simulation
+                new_solver->setSensitivityMethod(SensitivityMethod::forward);
+            } else {
+                new_solver->setSensitivityMethod(SensitivityMethod::none);
+                new_solver->setSensitivityOrder(SensitivityOrder::none);
+            }
+            new_solver->setup(t0, &model, ws_->x, ws_->dx, ws_->sx, ws_->sdx);
+            preeq_solver_unique_ptr_ = std::move(new_solver);
+            solver_ = preeq_solver_unique_ptr_.get();
+        }
+
+        runSteadystateSimulationFwd(model);
+
+        return SteadyStateStatus::success;
+    } catch (IntegrationFailure const& ex) {
+        switch (ex.error_code) {
+        case AMICI_TOO_MUCH_WORK:
+            if (model.logger)
+                model.logger->log(
+                    LogSeverity::debug, "EQUILIBRATION_FAILURE",
+                    "AMICI equilibration exceeded maximum number of"
+                    " integration steps at t=%g.",
+                    ex.time
+                );
+            return SteadyStateStatus::failed_convergence;
+        case AMICI_RHSFUNC_FAIL:
+            if (model.logger)
+                model.logger->log(
+                    LogSeverity::debug, "EQUILIBRATION_FAILURE",
+                    "AMICI equilibration was stopped after exceedingly"
+                    " long simulation time at t=%g.",
+                    ex.time
+                );
+            return SteadyStateStatus::failed_too_long_simulation;
+        default:
+            if (model.logger)
+                model.logger->log(
+                    LogSeverity::debug, "OTHER",
+                    "AMICI equilibration failed at t=%g.", ex.time
+                );
+            return SteadyStateStatus::failed;
+        }
+    } catch (AmiException const& ex) {
+        if (model.logger)
+            model.logger->log(
+                LogSeverity::debug, "OTHER", "AMICI equilibration failed: %s",
+                ex.what()
+            );
+        return SteadyStateStatus::failed;
+    }
+}
+
+[[noreturn]] void SteadystateProblem::handleSteadyStateFailure(
+    bool tried_newton_1, bool tried_simulation, bool tried_newton_2
+) const {
+    // Throw error message according to error codes
+    std::string errorString = "Steady state computation failed.";
+    if (tried_newton_1) {
+        errorString.append(" First run of Newton solver failed");
+        writeErrorString(errorString, steady_state_status_[0]);
+    }
+    if (tried_simulation) {
+        errorString.append(" Simulation to steady state failed");
+        writeErrorString(errorString, steady_state_status_[1]);
+    }
+    if (tried_newton_2) {
+        errorString.append(" Second run of Newton solver failed");
+        writeErrorString(errorString, steady_state_status_[2]);
+    }
+    throw AmiException(errorString.c_str());
+}
+
+realtype
+SteadystateProblem::getWrmsFSA(Model& model, WRMSComputer& wrms_computer_sx) {
+    // Forward sensitivities: Compute weighted error norm for their RHS
+    realtype wrms = 0.0;
+
+    // we don't need to call prepareLinearSystem in this function, since it was
+    // already computed in the preceding getWrms call and both equations have
+    // the same Jacobian.
+
+    xdot_updated_ = false;
+    for (int ip = 0; ip < model.nplist(); ++ip) {
+        model.fsxdot(
+            ws_->t, ws_->x, ws_->dx, ip, ws_->sx[ip], ws_->dx, ws_->xdot
+        );
+        if (newton_step_conv_) {
+            newton_solver_.solveLinearSystem(ws_->xdot);
+        }
+        wrms = wrms_computer_sx.wrms(ws_->xdot, ws_->sx[ip]);
+        // ideally this function would report the maximum of all wrms over
+        // all ip, but for practical purposes we can just report the wrms for
+        // the first ip where we know that the convergence threshold is not
+        // satisfied.
+        if (wrms > conv_thresh)
+            break;
+    }
+    // Just report the parameter for the last `ip`. The value doesn't matter -
+    // it's only important that all of them satisfy the convergence threshold.
+    return wrms;
+}
+
+bool SteadystateProblem::checkSteadyStateSuccess() const {
+    // Did one of the attempts yield a steady state?
+    return std::ranges::any_of(
+        steady_state_status_, [](SteadyStateStatus status
+                              ) { return status == SteadyStateStatus::success; }
+    );
+}
+
+void SteadystateProblem::runSteadystateSimulationFwd(Model& model) {
+    if (model.nx_solver == 0)
+        return;
+
+    // Do we also have to check for convergence of sensitivities?
+    auto sensitivity_method = SensitivityMethod::none;
+    if (solver_->getSensitivityOrder() > SensitivityOrder::none
+        && solver_->getSensitivityMethod() == SensitivityMethod::forward) {
+        sensitivity_method = SensitivityMethod::forward;
+    }
+    // If forward sensitivity computation by simulation is disabled,
+    // disable forward sensitivity integration in the solver.
+    // Sensitivities will be computed by newtonsolver.computeNewtonSensis then.
+    if (model.getSteadyStateSensitivityMode()
+        == SteadyStateSensitivityMode::newtonOnly) {
+        solver_->switchForwardSensisOff();
+        sensitivity_method = SensitivityMethod::none;
+    }
+
+    // function for sensitivity convergence check or dummy
+    std::function<bool()> sensi_converged;
+    if (solver_->getSensiSteadyStateCheck()
+        && sensitivity_method == SensitivityMethod::forward) {
+        sensi_converged
+            = [&, wrms_computer_sx = WRMSComputer(
+                      model.nx_solver, solver_->getSunContext(),
+                      solver_->getAbsoluteToleranceSteadyStateSensi(),
+                      solver_->getRelativeToleranceSteadyStateSensi(),
+                      AmiVector(
+                          model.get_steadystate_mask(), solver_->getSunContext()
+                      )
+                  )]() mutable -> bool {
+            updateSensiSimulation();
+            // getWrms needs to be called before getWrmsFSA
+            // such that the linear system is prepared for newton-type
+            // convergence check
+            return getWrmsFSA(model, wrms_computer_sx) < conv_thresh;
+        };
+    } else {
+        sensi_converged = []() { return true; };
+    }
+
+    // Returns the WRMS for the current state
+    auto get_wrms_state = [&]() {
+        if (newton_step_conv_) {
+            newtons_method_.compute_step(ws_->xdot, {ws_->t, ws_->x, ws_->dx});
+            return wrms_computer_x_.wrms(newtons_method_.get_delta(), ws_->x);
+        }
+
+        return wrms_computer_x_.wrms(ws_->xdot, ws_->x);
+    };
+
+    // Checks for convergence of the steady state solution.
+    // Checks state and delta/xdot depending on the options.
+    auto check_convergence = [&](bool state_changed) {
+        if (state_changed) {
+            // update current solution
+            flagUpdatedState();
+            updateRightHandSide(model);
+        }
+        wrms_ = get_wrms_state();
+        if (wrms_ < conv_thresh && sensi_converged()) {
+            return true;
+        }
+        return false;
+    };
+
+    updateRightHandSide(model);
+
+    EventHandlingSimulator sim(&model, solver_, ws_, nullptr);
+
+    int const convergence_check_frequency = newton_step_conv_ ? 25 : 1;
+
+    sim.run_steady_state(
+        check_convergence, convergence_check_frequency, numsteps_.at(1)
+    );
+
+    if (numsteps_.at(1)) {
+        flagUpdatedState();
+    }
+
+    // if check_sensi_conv_ is deactivated, we still have to update sensis
+    if (sensitivity_method == SensitivityMethod::forward)
+        updateSensiSimulation();
+}
+
+void SteadystateProblem::findSteadyState(Model& model, int it, realtype t0) {
+    steady_state_status_.resize(3, SteadyStateStatus::not_run);
+    // Turn off Newton's method if 'integrationOnly' approach is chosen for
+    // steady-state computation or newton_maxsteps is set to 0 or
+    // if 'integrationOnly' approach is chosen for sensitivity computation
+    // in combination with forward sensitivities approach. The latter is
+    // necessary as numerical integration of the model ODEs and corresponding
+    // forward sensitivities ODEs is coupled. If 'integrationOnly' approach is
+    // chosen for sensitivity computation it is enforced that steady state is
+    // computed only by numerical integration as well.
+    bool const turnOffNewton
+        = model.getSteadyStateComputationMode()
+              == SteadyStateComputationMode::integrationOnly
+          || solver_->getNewtonMaxSteps() == 0
+          || (solver_->getSensitivityOrder() >= SensitivityOrder::first
+              && model.getSteadyStateSensitivityMode()
+                     == SteadyStateSensitivityMode::integrationOnly
+              && ((it == -1
+                   && solver_->getSensitivityMethodPreequilibration()
+                          == SensitivityMethod::forward)
+                  || solver_->getSensitivityMethod()
+                         == SensitivityMethod::forward));
+
+    bool const turnOffSimulation = model.getSteadyStateComputationMode()
+                                   == SteadyStateComputationMode::newtonOnly;
+
+    // First, try to run the Newton solver.
+    if (!turnOffNewton)
+        findSteadyStateByNewtonsMethod(model, false);
+
+    // Newton solver didn't work, so try to simulate to steady state.
+    if (!turnOffSimulation && !checkSteadyStateSuccess())
+        steady_state_status_[1] = findSteadyStateBySimulation(model, it, t0);
+
+    /* Simulation didn't work, retry the Newton solver from last sim state. */
+    if (!turnOffNewton && !turnOffSimulation && !checkSteadyStateSuccess())
+        findSteadyStateByNewtonsMethod(model, true);
+
+    // Nothing worked, throw an as informative error as possible.
+    if (!checkSteadyStateSuccess())
+        handleSteadyStateFailure(
+            !turnOffNewton, !turnOffSimulation,
+            !turnOffNewton && !turnOffSimulation
+        );
+}
+
+void SteadystateProblem::findSteadyStateByNewtonsMethod(
+    Model& model, bool newton_retry
+) {
+    int const stage = newton_retry ? 2 : 0;
+    try {
+        updateRightHandSide(model);
+        newtons_method_.run(
+            ws_->xdot, {ws_->t, ws_->x, ws_->dx}, wrms_computer_x_
+        );
+        steady_state_status_[stage] = SteadyStateStatus::success;
+    } catch (NewtonFailure const& ex) {
+        switch (ex.error_code) {
+        case AMICI_TOO_MUCH_WORK:
+            steady_state_status_[stage] = SteadyStateStatus::failed_convergence;
+            break;
+        case AMICI_NO_STEADY_STATE:
+            steady_state_status_[stage]
+                = SteadyStateStatus::failed_too_long_simulation;
+            break;
+        case AMICI_SINGULAR_JACOBIAN:
+            steady_state_status_[stage]
+                = SteadyStateStatus::failed_factorization;
+            break;
+        case AMICI_DAMPING_FACTOR_ERROR:
+            steady_state_status_[stage] = SteadyStateStatus::failed_damping;
+            break;
+        default:
+            steady_state_status_[stage] = SteadyStateStatus::failed;
+            break;
+        }
+    }
+    numsteps_.at(stage) = newtons_method_.get_num_steps();
+    wrms_ = newtons_method_.get_wrms();
+    flagUpdatedState();
+}
+
+void SteadystateProblem::flagUpdatedState() {
+    xdot_updated_ = false;
+    sensis_updated_ = false;
+}
+
+void SteadystateProblem::updateSensiSimulation() {
+    if (sensis_updated_)
+        return;
+    ws_->sx = solver_->getStateSensitivity(ws_->t);
+    sensis_updated_ = true;
+}
+
+void SteadystateProblem::updateRightHandSide(Model& model) {
+    if (xdot_updated_)
+        return;
+    model.fxdot(ws_->t, ws_->x, ws_->dx, ws_->xdot);
+    xdot_updated_ = true;
 }
 
 } // namespace amici
